Correctly decoding phase information
I have decided to encode the phase information in the 32:nd bit of the k
index in the index triples. This is normally how signed integers work,
however, the sign is not used for zero. I needed to add sign information
even for the zero value which is done with some bit wise operations.