Added the code that creates binary index lists
The anicr code produces index lists in a human readable format. This is
obviously slow to read and Minerva has to do it many times. Aurora
transforms the index lists to a binary format. Furthermore, anicr is not
aware of the array index that each array is given, it is generated in a
the pearlscript that creates the comb.txt files.