Renamed transformed_block_manager
It has been clear that the best is to separate the 2nf handling
completely from the 3nf handling. Therefore, to not let their namespaces
collide, and to make the difference more clear I had to rename the
transformed_block_manager to transform_2nf_block_manager, since it only
handles, and should only handle, 2nf.