checkout global_constants from master
In the master-branch we had fixed an issue with older versions of
gcc, they for some reason didn't consider const declared global
variables as constants and it was therefore not possible to use them
in global constant expressions. Newer versions has no problem with this.
We solved this issue by having macros for older versions and constants
for newer.