Parallel main structure written
The main structure of the new parallel version of Mercury's 3nf code has been
written. It is based on the idea that all blocks within the same energy block
is computed in sequence while the energy blocks is supposed to be computed in
parallel. This is because the blocks within an energy block share data while
blocks in different energy blocks do not.