Fetch execution instruction lock
To make sure that no two threads fetches the same instruction it is
necessary to use a omp_lock_t. The lock is set by the
has_next_instruction, if there is no next instruction
has_next_instruction unlocks it again. However, if there is a next
instruction, to make sure that no other thread fetches that instruction
has_next_instruction leaves the lock locked and expects that the user
calls next_instruction which then will unlock the lock.