First prototype of multi threaded mat vec mult
The first prototype of the multi threaded matrix vector multiplication
needed for Lanczos has been completed. However, the system for loading
and unloading data has some problems. If an unloading block is initiated
before the arrays it should unload has not been loaded yet the code
emits an error. It is prudent to make sure that unloading blocks are
only called after the corresponding arrays has been fully utilized.

Furthermore, this prototype would give erroneous result in the
multi threaded case (OMP_NUM_THREADS > 1), since the output vectors are
not handled correctly, there should be one array per thread, that is
reduced to a single array when writing to file.