Any thing before BLOCK and UNLOAD_BLOCK is ignored
A problem with greedyorder is that it sometimes output information on
how much of the calculation is left in to the execution order list. This
output is always in the beginning of the row, and the useful information
is outputed after it. Therefore, execution order now ignores everything
before the UNLOAD_BLOCK and BLOCK.