Restructuring the memory management system
The new memory management system is based on having one thread only
tasked with loading and unloading memory, while the others only perform
matrix vector multiplications. At first it loads memory until it can't
load any more, and it loads it in the order that is needed by the
threads. As it needs to load more memory it waits until there exists
arrays that are currently not in use, and will not be needed by any of
the calculation blocks between the leading multiplication thread and the
memory management thread.