Fixed a bug
For some reason, the row that resets the current single particle basis
to NULL was outside the critical block. In large enough model spaces it
is possible that this variable was reset by a different thread while the
current thread was inside the qsort in the critical block thus causing
segmentation fault. This bug was solved by moving the reset row inside
the critical block