Fixed some final bugs in Vulcan
If a matrix block file does not exist, Vulcan should treat it as being
identical to zero. Also, if a matrix block only is for interactions
between protons or neutrons, the length of the block was wrong, but now
it is computed correctly. I had for some reason flipped the input buffer
with the output buffer.