Fixed a bug
As I was running the ⁶Li Nmax=10 case on one tetralith:s smaller nodes I
discovered that bacchus randomly got stuck in an infinite loop after
about 20 Lanczos iterations. After replicating the bug on my machine, by
heavily restrict memory and rerunning ⁴He Nmax=2 case multiple times I
realized that it was due to sloppy use of a omp_nest_lock_t that was used to
1) for load_array to check if an other load_array was loading the same
array
2) for is_array_loaded to check if load_array was loading that
particular array.
Since is_array_loaded was used in make_space_for to check if a thead, A,
could safely unload an array, k, there where a tiny probablity that this check
occured just before a load_array in an other thread, B,  check if array, k, are
loading, and thus believing that it was already loading in an other thread.

The consequence of this was that as thread B enter a request for array k,
array k was not loaded and will not be loaded so thread B get locked.
Then if there is not enough memory for thread A to unload A would get in
an infinite loop.

All the circumstances needed to trigger this bug is very unlikely to occur,
and can only occur when there are strict memory restrictions. That is
why I haven't seen it before now.

I solved this bug by introducing a new omp_lock_t, that was only used
for case 1), and therefore it is impossible for the situation above to
occur.