Fixed a bug
The target_eigenvector should be indexed from 0 and not 1. It is used in
two different contexts, determining which eigenvector to apply the
convergence criteria and, which to save. However, in the former case it
is assumed to be indexed from zero, but in the latter case it is assumed
to be indexed from one. I have now changed that so that it is assumed to
be indexed from zero in both cases.